Move filesystemDetect function into lxd/util subpackage.
This will make it possible to use util.FilesystemDetect from other
sub-packages, in particular the lxd/sys one.
