+++ lxd/util/fs_32bit.go
@@ -1,9 +1,9 @@
 // +build 386 arm ppc s390
 
-package main
+package util
 
 const (
 	/* This is really 0x9123683E, go wants us to give it in signed form
 	 * since we use it as a signed constant. */
-	filesystemSuperMagicBtrfs = -1859950530
+	FilesystemSuperMagicBtrfs = -1859950530
 )
