+++ lxd/util/fs_64bit.go
@@ -1,7 +1,7 @@
 // +build amd64 ppc64 ppc64le arm64 s390x
 
-package main
+package util
 
 const (
-	filesystemSuperMagicBtrfs = 0x9123683E
+	FilesystemSuperMagicBtrfs = 0x9123683E
 )
